extract social and entities query building
- from query builder to type search controller
- reuse social query builder in user search
- start dropping body_builder naming
- add some comments

since user and group have different attributes (username vs name) there
is no need to filter query by types, indexation of a type for those
documents become unecessary but its okay to leave type ikey as it
doesnt cost much to leave it and it could be useful in the future.
