+++ server/controllers/search/lib/type_search.js
@@ -5,7 +5,8 @@ const assert_ = __.require('utils', 'assert_types')
 const { host: elasticHost } = CONFIG.elasticsearch
 const { formatError } = __.require('lib', 'elasticsearch')
 const getIndexesAndTypes = require('./get_indexes_and_types')
-const queryBodyBuilder = require('./query_body_builder')
+const entitiesQueryBuilder = require('./entities_query_builder')
+const socialQueryBuilder = require('./social_query_builder')
 
 // types should be a subset of ./types possibleTypes
 module.exports = ({ lang, types, search, limit = 20 }) => {
@@ -16,8 +17,16 @@ module.exports = ({ lang, types, search, limit = 20 }) => {
   ({ indexes, types } = getIndexesAndTypes(types))
 
   const url = `${elasticHost}/${indexes.join(',')}/_search`
+  let body = {}
 
-  const body = queryBodyBuilder({ lang, types, search, limit })
+  // query must be either social (user, group) or entities related
+  // but cannot be both as results scores are built very differently
+  if (types.includes('user') || types.includes('group')) {
+    body = socialQueryBuilder(search)
+  } else {
+    body = entitiesQueryBuilder({ lang, types, search, limit })
+  }
+
   return requests_.post(url, { body })
   .catch(formatError)
 }
